fix: change price_per_sqm to BigInt and add serialization
Fix integer overflow when migrating price_per_sqm from Decimal. Use BigInt
column type and add Number() conversion in formatRow for JSON serialization.

+++ apps/backend/prisma/migrations/20260412010000_change_price_per_sqm_to_int/migration.sql
@@ -1,2 +1,2 @@
--- Round existing values and change column type to integer
-ALTER TABLE "properties" ALTER COLUMN "price_per_sqm" TYPE INTEGER USING ROUND("price_per_sqm")::INTEGER;
+-- Round existing values and change column type to bigint
+ALTER TABLE "properties" ALTER COLUMN "price_per_sqm" TYPE BIGINT USING ROUND("price_per_sqm")::BIGINT;
